Fix pvr sync fence file leak
BZ: 111274

In PVR KM to free the fence ref from a fence array, it will break the
loop if it encounter a null fence in it. But for video overlay case,
video driver or camera will not provide a fence for the overlay buffer
and a null fence object will returned by pvr sync driver. So if video
is first buffer and another rgb buffer is in next slot. the rgb buffer
ref will never be decreased. It will run out of the kernel file struct
pool after a long time video or camera test.

Change-Id: I528b4d4f4f31672a537b7663230909ba3b90834e
Signed-off-by: Marc Mao <marc.mao@intel.com>
Reviewed-on: http://android.intel.com:8080/115510
Reviewed-by: Liu, MinX Z <minx.z.liu@intel.com>
Tested-by: Liu, MinX Z <minx.z.liu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>