MRFLD: HDMI: Address the issue that HDMI monitor is randomly blank after hot-plug
BZ: 119493

Previously in intel_output_poll_execute(), a local variable old_status
is used to save last HDMI connection status via copying the value of
connector->status. However connector->status may be modified asynchronously
in other threads even before the function gets called, which makes it unsafe.
So instead of using old_status, we use the member variable "encoder" in
DRM_CONNECTOR to check if the HDMI connector is enabled or not.

Change-Id: Icab172a74163c3c28fe261b4e1d7f542122aa8f9
Signed-off-by: Jason Chen <jason.chen@intel.com>
Reviewed-on: http://android.intel.com:8080/116022
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Hu, Austin <austin.hu@intel.com>
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Tested-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>