RGX: Use memory pools instead of vmaps
BZ: 114257

When using vm_map_ram with small blocks, blocks of 2MB
are created to avoid fragmentation in vmalloc
These blocks are cleaned when no allocations remains
to avoid too many tlb flushs
This behavior leads to a creation of a lot of vmap_blocks
of 2MB and quickly fill the vmalloc area because some
rgx allocations can remain a long time
Using memory pools in the vmalloc area for rgx avoid the
new_vmap_block issue

IMG Ticket# 26118
IMG did not accept all our changes. A patch for osfunc.c
was delivered. But, it will not be merged to IMG's
mainline. According to the patch, physmem_osmem_linux.c
was modified.

Change-Id: Iba8b57d4b02b319a80b2f608a826961d4d284e23
Signed-off-by: Philip Hatcher <philip.hatcher@intel.com>
Signed-off-by: Julien Delayen <julienx.delayen@intel.com>
Reviewed-on: http://android.intel.com:8080/118023
Reviewed-by: Woo, Insoo <insoo.woo@intel.com>
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Tested-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>