MRFLD: OSPM: refined the PCI resume/suspend in power_island_get()/put().
BZ: 123411

We should only resume/suspend the registered PCI in power_island_get/put,
by: resuming PCI at the beginning of power_island_get(), if all of the 12
islands are off; suspending it at the end of power_island_put(), if all
of the 12 islands are off.

And we should init display firstly as it keeps DPA island on, otherwise
get/put VED island in psb_do_init() would suspend PCI so that the panel
turned on by firmware will be black.

Conflicts:

	display/tng/drv/ospm/pwr_mgmt.c

Change-Id: I26e75e681695516441e9a157b86009d62d81b77e
Signed-off-by: Austin Hu <austin.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/118598
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Tested-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>