[REVERTME] MRFL: GFX-Display: ported the changes from patch #117334.
BZ: 110361

Not to do page flip if the system is suspended, to avoid the ipanic issue
reported in BZ #120960, without the lock protection which conflicts
with patch #120040 (lock confliction between page flipping and vsync
waiting).

Need to check whether the patch could be reverted, with the OSPM refined
patches (#118597 & #118598) applied.

Change-Id: I5d32f7c43528d36be4fdc4a7ec6abe730968e387
Signed-off-by: Austin Hu <austin.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/120055
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Tested-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>