[Video] Implement VSP secure firmware uploading for B0
BZ: 126241

B0 is a new stepping for Merrifield. Compared with A0, it added
secure video playback feature. This feature requires VSP to execute
a verified firmware and do post-processing with IED encrypted data.
In order to implemented this feature, this patch followed a well
defined boot flow to copy firmware to IMR, and use chaabi FW verify
and setup the protection for the data. Then get PUnit upload the
firmware correctly.

Signed-off-by: Binglin Chen <binglin.chen@intel.com>

Change-Id: I46be999de839f7f485b53bbf69ba62ac55b1f146
Reviewed-on: http://android.intel.com:8080/120673
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>