Optimize if-statements while avoiding a Klocwork false-positive
BZ: 133609

Avoid a Klocwork false positive while at the same time optimizing
some if-statements.

There existed an if-statement:
    if (cmd & (IOC_IN | IOC_OUT))
that was immediately followed by:
    if (cmd & IOC_IN) ...
    else if (cmd & IOC_OUT) ...

The two follow-on if-statements can be included within the first
if-statement.  This results in better code and should avoid a Klocwork
false positive warning about use of a variable initialized under
control of the first if-statement within the later if-statements.

Change-Id: I1e2b854b3d2030119685abe532f5bf9bdd8846fa
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/127766
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: Poornachandran, Rajesh <rajesh.poornachandran@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>