MRFLD: Display: improve panel reset procedure to workround TE loss
BZ: 118214

we submit this new patch to assist patch 127621. when TE sometimes loss,
we need to reset the panel, but for now the reset procesure is not
enough in case the display island  was just powered off then the pipstat
will loss and reset to default, and TE will no longer come even we reset
the panel, so this patch improve it by recovering the pipestat when needed.

Change-Id: I95be2d097f968ef2068bb4119453027c4fada1c8
Signed-off-by: wang, biao <biao.wang@intel.com>
Signed-off-by: Zhang, Di <di.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/128469
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Tested-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>