GFX-PM: [REVERT-ME] Workaround for Blocking S0i3 when Display is ON
BZ: 142445

When all graphics and display islands are in D0i3 (DSR+APM)
system will try to go into S0i3 when screen is ON.  But
Platform will hang.  We need to block S0i3 when display is ON.

Revert the workaround when platform hang is root caused and fixed.

Change-Id: Ida255324f19e77ea11f07da7e64ac4a202c8065f
Signed-off-by: Hitesh K. Patel <hitesh.k.patel@intel.com>
Reviewed-on: http://android.intel.com:8080/135703
Reviewed-by: Salah, Haithem <haithem.salah@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>