MRFLD: Display: fixed the issue that vsync enabling randomly fails during page flip.
BZ: 116643

The issue can be easily reproduced by decreasing DSR count to 1 and running
gears with HDMI connected.

The root cause is that DSR allow is invoked during the page flip process,
rather than in the vsync interrupt MISR, so that DSR happens and invokes
drm_vblank_off() at the next TE, but drm_vblank_put() will be invoked at the
TE(s) later because we need to hold the current displaying buffer. So,
drm_vblank_get() fails with vblank_refcount greater than 0 and vblank_enabled
set as 0.

Change-Id: I9360ee9955b2ac37d4f10396bef9a04ab73fdad0
Signed-off-by: Austin Hu <austin.hu@intel.com>