MRFLD: Display: fixed another issue that vsync enabling randomly fails during page flip.
BZ: 116643

The issue could be reproduced by DPMS on/off MIPI with 3D app (gears)
running with stress test.

The root cause is that sometimes the clocks from DSI PLL are not locked
to Pipe A, so that enabling TE interrupt would fail even though the
power on sequence has completed. Fixed the problem by waiting DSI PLL to
be locked during the after enabling it (only need to poll locking status
of pipe A, as both MIPI pipes share the same DSI PLL).

Change-Id: I46f516b2bdfb14e7732489e869589fcf0f0d061b
Signed-off-by: Austin Hu <austin.hu@intel.com>