[PORT FROM MAIN JB-MR2] [VIDEO] make sure pci is powered on before writing GUNIT address
BZ: 149147

FE happened when psb was wrting a GUNIT address, while the PCI is
not resumed completely. At that moment, the mmadr is 0x0, which will
indicate an error of address hole.

This patch fixes this issue by get_sync/put_sync the pci runtime power
management before writing the GUNIT adrress.

Change-Id: I1206ce25d6dd92fec9e16f35d7587ee7e565aeab
Orig-Change-Id: I6f122fd9661912db2c9219f0a0011ff005f710a5
Signed-off-by: Zhi Jin <zhi.jin@intel.com>
Signed-off-by: CHEN Gang <gang.c.chen@intel.com>