gfx: Enabling RI_DEBUG in debug build
BZ: 159033

kernel config variable named CONFIG_RI_DEBUG is defined
in gfx kernel makefile. But, it is not defiend in
kernel config file.
And PVR_RI_DEBUG macro needs to be added as a compile
option.
The compile option mismatch between user and kernel modrivers
causes failure to boot to UI in debug build.

Change-Id: I96d2efb8dd98b60b14ad982939ff9d834c958ac5
Signed-off-by: insoowoo <insoo.woo@intel.com>