[Graphics]: Fix the bUnsetMemoryType flag set logic erro in Phymem managemant
BZ: 149534

There are two errors in the bUnsetMemoryType flag set logic:
1. When open mem pool, it forget to set the flag when get pages from mem pool,
so when free pages, it will not set to WB flag back.
2. Even not open mem pool,When alloc pages array, if the last index page alloc failed,
it will set the flag to False, but the former pages are alloced into WC or UC type.
So it will not free the former pages back to WB type.
These are two root cause that a lot of os pages are changed into WC or UC type and never
changed back to WB type.
When kernel malloc some data structure on these WC or UC type memory.
The data can't be protect by mutex or spinlock well.
So when multiple CPU access the same data into these WC or UC type memory,
it will crash abnormally.

Change-Id: I80ab28771cac829bda02cc01a461711af9d782f9
Signed-off-by: Shi Yang <yang.a.shi@intel.com>
Signed-off-by: Ma Xindong <xindong.ma@intel.com>
Signed-off-by: Nie Jun <jun.nie@intel.com>