[DEVFREQ]: need to check if init is done in df_rgx_bus_target()
BZ: 164484

It is possible that df_rgx_bus_target() is running but the module loading
hasn't been finished yet. So it could cause IPANIC. Here is the case in the
bug.

init thread                               kworker thread
df_rgx_busfreq_probe
->devfreq_add_device
  ->devfreq_monitor_start
    //the work is queued
...
->thermal_cooling_device_register
  -->sysfs_addrm_start
     ->mutex_lock_nested                 devfreq_monitor
                                         ->update_devfreq
				           ->df_rgx_bus_target
                                             ->mutex_lock(g_mutex_sts)
                                             //this mutex hasn't been init yet

...
->dfrgx_burst_init
  ->mutex_init(g_mutex_sts)
    //mutex is init here

So add a check of init done in df_rgx_bus_target to avoid it.

Change-Id: Ia0b9e6cab51c6d549a47cf1d3b9a6bfbb170c421
Signed-off-by: Zhang Di <di.zhang@intel.com>