Display: change flip retire strategy in display IRQ
BZ:156311

This patch changes display flip retire strategy when IRQ comes.
It is only safe to release previous flip when next flip comes and
updated on screen with vsync comes after that.
The patch is used to fix the tearing on HDMI when playing video
with sync mode while only one buffer canceled from video.

Change-Id: I6e96a6fbeaf3915d77e6a816c30e3f54e1c2ddf6
Signed-off-by: Lei Zhang <lei.zhang@intel.com>