GFX - Enable RGX IRQs when power island turned on
BZ: 166169

This is the same change as http://android.intel.com:8080/#/c/140698

During ANN power on, we disabled the code where we enable IRQs for
the RGX island. This change just enables the correct code that was
commented out incorrectly.

Basically, we enable RGX irqs after we turn on the RSCD islands.

Change-Id: I09e39d77cc2e04722094f4138401e6e5ba963285
Signed-off-by: Vinil Cheeramvelil <vinil.cheeramvelil@intel.com>
Signed-off-by: Austin Hu <austin.hu@intel.com>