Use tasklet feature to report HDMI audio buffer done event to HDMI audio driver
BZ: 164563

In normal scenario the a buffer is rendered in 21ms. But during the error scenario the buffer done
interrupt is received by audio driver after ~200ms and by that time other 3 buffers are rendered
by hw and empty. This results in the audio loss.

When receiving HDMI_AUDIO_BUFFER_DONE interrupt,  the graphic interrupt handler schedule a work queue to
report the HDMI_AUDIO_BUFFER_DONE event to HDMI audio driver, that will casue the "~200ms" delay

The HDMI_AUDIO_BUFFER_DONE event should be reported to HDMI audio driver as soon as possible, in Linux
kernel the tasklets run in software interrupt, it's quicker than work queue

Change-Id: Id35c8758d8b0dc4719b1e5a686f150df9cc51a76
Signed-off-by: lychenx <liy.x.chen@intel.com>