hdmi-audio: Fix error handling on bad probe sequence
BZ: 168420

There are some error handling paths that are incorrectly handled,
specially when mid_hdmi_audio_setup fails, which goes to an error
handling section which assumes intelhaddata->dev has an actual
struct device pointer, but that doesn't happen until later on
in the init.

Clean up the sequence and fix this potential null pointer
dereference.

Change-Id: Ia6a4752e5775ede8d0c136b58cef48bef0a49bdb
Signed-off-by: Sergio Aguirre <sergio.a.aguirre.rodriguez@intel.com>