GFX-Display: fastboot UI support on big panel
BZ: 173232

Allocate psb framebuffer according to different Panel used.
This patch needs IA FW to reserve different stolen memory size according to the panel used.

Now because we don't have enough stolen memory, the driver allocate a fixed framebuffer.
This is the reason that the fastboot UI didn't display on the Sharp 1080P and 25x16 Panel.

Change-Id: Ifde1a4bf5dadfefb150c17fd1bd9b2896fa4b409
Signed-off-by: Liping Shan <liping.shan@intel.com>