GFX: MRFL/MOFD - Move DSI PLL lock to before MIO power up
BZ: 174801

There are some situations where the DSI link is not properly
established. The DSI PLL locking has been moved prior to
MIO power on in order to avoid race condition of DSI PLL lock
failing and leading to vsync and other DBI errors. Refer to
BZ and HSD for details on this issue

Change-Id: If60be939c44877d8a793a96ede6c4b9953b79357
Signed-off-by: Vinil Cheeramvelil <vinil.cheeramvelil@intel.com>