[MOFD_V1][Video_Edit]TOPAZ time out during exporting clip form movie studio
BZ: 176769

remove the error check for tng_topaz_set_vec_freq error in vec power function.
tng_topaz_set_vec_freq function only adjusts the VEC frequency. when it occurs error, the VEC frequency would remain the old one.
this kind of error is not fatal for driver and system, and also is only common on VV board.
so VEC driver could check this error, not halt by it.

Change-Id: Ib2123be921fdb777007209d2c4003d0d92f55155
Signed-off-by: edward lin <edward.lin@intel.com>