Display: set plane state to be correct in PowerOff
BZ:185181

This is kernel patch to set plane state to be correct in PowerOff.
It has a corner case in hdmi video playing while doing plug out
operation, where it may cause the disable state has no chance to
be changed correctly in ISR handler, and then cause HWC query on
plane state is wrong and no chance to reset plane.

Change-Id: I371717b2e27a458aba06e1fd034136e345f8a9cd
Signed-off-by: Lei Zhang <lei.zhang@intel.com>