psb_fb: fix garbage on droidboot when connect HDMI
BZ: 186680

when HDMI connected, sizes->fb_width/fb_height will be
the min width/height between the HDMI and MIPI, this
may lead the wrong setting of screen's resolution.

Now adopt the CTP's way: allocate fb which size is same as MIPI.
on HDMI side, it will scalling the fb's content to match the
HDMI screen.

Change-Id: I79cfe35c1592627b0ba0017ae49e0408f695bb17
Signed-off-by: Tan Wei <weix.z.tan@intel.com>