dc: fix flip context corruption
BZ: 193971

Now we post a flip request from HWC and find the according buffers
by the handle in kernel and copy user's context to its own context,
finally submit a scp command and let wq to run the command.

This model exists a problem: if a new flip request comes just after
submiting command and before running the command, moreover, this
flip request contains some same buffers within previous flip request.
It will overwrte the buffer's context because we reset the context's
count before setting. Bang!!!

To prevent this happens, we allocate a copy of the buffer when
setting the context rather than doing this on _Queue_Flip.

Change-Id: I6419e098f1e0d0eccfc6d3d22befd04383fd3571
Signed-off-by: Tan Wei <weix.z.tan@intel.com>