GFX-Display: fix loading tngdisp module for a long time
BZ: 195647

Patch#184686 exposed that kernel will be trapped in load tngdisp module.
The root cause is that we enable the TE too ealier. It takes a long time
to handle the TE interrupt during init, which will block the whole
module init. And kernel watchdog can't be launched and kicked. Then the
crash occurred.
This patch moved the TE at the end of tngdisp driver load.

Change-Id: I497e35c6da36044e3ee00f4510c45f5dac6dfa1e
Signed-off-by: wang, biao <biao.wang@intel.com>
Signed-off-by: Zhang Di <di.zhang@intel.com>
Signed-off-by: CHEN Gang <gang.c.chen@intel.com>