dc: fix buffer memory leak when configure failure
BZ: 194945

When hwc posts flip request, sometimes it will failed because of no enough scp
command space and some other reasons, it will retry until it succeed. Because we
copy dc_buffer in configure check, so it will allocate a copy everytime, but
only the last one is used.

Now we allocate buffer copy when everything is done.

Change-Id: I68850b48c1ce9c4ed6879cbf25d191c396066220
Signed-off-by: Tan Wei <weix.z.tan@intel.com>