Display: Enable high memory support for 32 bits
BZ: 200496

In 32 bits, low memory is no more than 900M. If display resolution
is high, low memory might be used up. This change is to enable
high memory for 32 bits. Another small change is to skip restoring
cache attribute for high memory because it is not saved during
memory allocation.

Change-Id: I349b3c5db609c00a60febab5413af6612023b375
Signed-off-by: Wayne Wu <wayne.wu@intel.com>
Signed-off-by: Junxiao Chang <junxiao.chang@intel.com>