tngdisp: fix screen not resumed in fastboot & recovery mode
BZ: 200493

We can't set 0 to DSPACNTR, which causes plane not enabled after power
off and on in fastboot and recovery mode. In fastboot & recovery mode,
DSPACNTR will not be touched after power on.

While in Android GUI, DSPACNTR will be updated by flip path, so no
problem to write 0 to it during power on.

Change-Id: I640009b44d7b0a2afc2175b30de65501b6d6b62c
Signed-off-by: Hong Liu <hong.liu@intel.com>