Display: add flip timer to monitor flip status
BZ: 203793

This patch adds flip timer as a watchdog to monitor flip status.
It will be setup everytime when there comes a new flip or there
comes a vsync ISR event. And it will run out of time if vsync
lost or long time there is no new flip comes.
In the timer handler, it will free all flip events on flipqueue.

The patch is targeted for checking and unblocking 3D fence blocked
issue from Display side. It has risk to cause tearing in some cases.

Change-Id: Ica03dbc187ae0ded03794bdb22f14d0d2855244f
Signed-off-by: Lei Zhang <lei.zhang@intel.com>