pvr_sync: prevent timeline free when traverse timeline list
BZ: 203792

Now we release lock when signal a timeline because it may lead
to a timeline destroy which needs to acquire this lock again.
But this will give a chance to destroy a timeline directly
when we travel the timeline list.

This patch is from the patch in IMG ticket 29527.

Change-Id: I750d06675bf466f1437383324c04bc8feed18c1e
Signed-off-by: Tan Wei <weix.z.tan@intel.com>