Services: Remove APM from MISR and create new workqueue dedicated to APM
BZ: 209193

This patch is from IMG to replace previous fix. It should co-work with
the reverted patch.

This change prevents contention for FW access between an APM request and
bridge originated calls:

Last FW operation occurs so it indicates idle, and sends an APM request
to
host. However on the host, another command is about to be sent to the
firmware to clean up the run lists etc... so it has the bridge lock.

Thus the MISR thread is blocked, waiting for the bridgelock by APM
request.
The host (having both bridge and powerlock) sends the command to the
firmware which gets completed immediately, and sends an irq back to the
host to say the sync update completed and to unblock the poll.

However, the MISR is already active and blocked by the APM request. So
it
prevents the poll from being unblocked. As soon as the poll completes a
single wait period(10ms on TC) it rechecks the sync value and sees it
has
completed. Unblocking the APM request and the now redundant sync IRQ.

This cycle continues for every bridge call, delaying shutdown time by
up to 50-60ms, during which the GPU is powered + idle.

Change-Id: Ia84d63473cb0f9e525edc2e9eccb8ebb860d8001
Signed-off-by: Tong, Bo <box.tong@intel.com>