[Graphics]: GFX PCI device need set master bit to enable it
BZ: 212817

When do kexec, the master bit of GFX pci device will be cleared.
And GFX driver never set it again when boot up from kexec.
So interrupt are not enabled for GFX pci device even if the display and
GPU are all powered on and can works well.

So fix it by enable this master bit when do psb_driver_load.

Change-Id: Iab69f5fbf981720a521e139141dce732e83ae828
Signed-off-by: He Bo <bo.he@intel.com>
Signed-off-by: Zhang Yanmin <yanmin.zhang@intel.com>
Signed-off-by: Shi Yang <yang.a.shi@intel.com>