vsync: delete unnecessary locking on wait vsync triggered by HWC
BZ: 211575

The locking just prevent dpms on/off or suspend/resume
happening when wait a vsync. But it isn't necessary.
Because when dpms off or suspend, it will wake up the
waiting thread and then disable the vsync.

The normal crtc mode set path will influence the vsync
wait because of this lock. This will lead some frames
stuck when hotplug hdmi.

Change-Id: I943217cd0e066174f95354ee79bb2aeb8ca553bc
Signed-off-by: Tan Wei <weix.z.tan@intel.com>