tngdisp: add support to skip panel init during first boot in command mode panel
BZ: 213476

To keep fw setting for the display controller and MIPI panel on first
boot in command mode panel, we need to keep display island on.

Panel wanting to skip driver modesetting on first boot needs to set
dsi_hw_context.panel_on to true in the panel_funcs->detect().

Change-Id: Ieea603420ddbd31823c88ae31093c07893547508
Signed-off-by: Ben Kao <ben.kao@intel.com>