dc: fix mipi panel can't resume issue
There are three problems lead pipe hang:
1. If we disable the last active plane on a enabled pipe,
pipe will hang.
2. When we enter external mode, we will call dpms standby,
and then we press the power button, mipi panel will be power off.
But hwc will call dpms on because of we exit external mode,
it will power on pipe again. When we press the power button to
turn on the panel, pipe will hang.
The solution is remove set power operation from dpms on,
dpms only control the panel's backlight.
3. There is a bug in DCCBEnablePrimaryWA: bit 31 won't be set
if we disabled primary before.

Change-Id: I0612ffd2b81b84c0cd361be101adf4ec9a1b72a6
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1938
Signed-off-by: Tan Wei <weix.z.tan@intel.com>
Reviewed-on: https://android.intel.com:443/260439