[Graphics]: protect the index exceed the max value
BZ: 213967

RGXProcessRequestFreelistsReconstruction only happened when RGX need to do HWR.
The count value are sent from RGX firmware to Kernel by CCB command directly.
We dount the HWR may caused by GPU memory corruption, and such corruption also
damage the memory content of CCB command.
And so the count value passed by CCB is invalid totally.
and then the index in the freelist array will be large than normal and the memroy
access will be invalid.

So add code to protect it.

Change-Id: Ife319cb1d9c3c2b3685178b45ad9cf38d7dc05e6
Signed-off-by: Shi Yang <yang.a.shi@intel.com>