pvr_sync: move back signaled flag back to PVR_SYNC_PT
BZ: 216415

The pt may be signaled when doing fence merge. The
signaled flag in PVR_SYNC_DATA will be updated
to 1. Then PVRSyncUpdateAllTimelines won't signal
any other pt that is a duplicate of this pt, since
they share the same signal flag. This function
think that such pts have been signaled before, and
there is no need to signal again. So we can see
two duplicated pts with different states.

Change-Id: I2780d2f4a5ed242efd8ae186706fd6d5112dbbc4
Signed-off-by: Jin Zhebin <zhebinx.jin@intel.com>