Return GPU Freq to Original Freq after Throttle
If the GPU frequency is throttled by PUNIT FW, the GPU frequency
will not return to the original minimum once throttling is done
and no longer needed... it remains at the throttled frequency.
This change corrects this behavior by trying to set the frequency
back to the original minimum as it is being throttled - thus, it
will return once the throttle is lifted.

Change-Id: I2fbe55b6fa5749317c5f27223c5e8d32cb9ddcd1
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1938
Signed-off-by: Robert M. Fosha <robert.m.fosha@intel.com>
Signed-off-by: Tan Wei <weix.z.tan@intel.com>
Reviewed-on: https://android.intel.com:443/260446