tngdisp: allocate framebuffer in 0~4G space
With 4G DDR memory installed on MOFD board, we can see 2G DDR memory is
located at [4G, 6G] memory space.

[    0.000000] BIOS-e820: [mem 0x0000000100000000-0x000000017fffffff] usable

So with current code, the framebuffer can be allocated with >4G physical
memory, this causes the entry we write to GTT is not right, since GTT
entry is 32bit only on MOFD.

Fixed by allocating FB with __GFP_DMA32 flag to keep it in 0~4G space.

Change-Id: I62ce90d0c935705b52178b7e07ec013f616af0b4
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-4418
Signed-off-by: Hong Liu <hong.liu@intel.com>
Reviewed-on: https://android.intel.com:443/279032
Reviewed-on: https://android.intel.com:443/285564