get vsync time in irq and not in work queue
The vsync is a time critical signal, its timestamp
is sent to the SurfaceFlinger that generates
several software VSYNC based on it before the next
synchronization.
That means the VSYNC timestamp reported by the kernel
to the user space should be as accurate as possible.

When vsync time is taken in the work queue there is
a delay between the interrupt and the workqueue.
From the measures I made during 10 s this delay
can be up to 3 ms, and is often several hundred
of us. This causes the SurfaceFlinger to drift on all
software VSYNC it generates so that on next
resynchronization with the hardware VSYNC it may
fire twice in less than 10 ms (I observed 5 ms).
This causes a UI jerk.

With this change the VSYNC time is taken in the interrupt
handler to be as closed as possible to the hardware
signal.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-8364
Change-Id: If720f9feb54c5b43a7bb933c75a4fcad776fce28
Signed-off-by: Thomas Buhot <thomas.buhot@intel.com>
Signed-off-by: Philip Hatcher <philip.hatcher@intel.com>
Reviewed-on: https://android.intel.com:443/298301